tweaking a few default vars; adding readme

+++ build.sh
@@ -38,7 +38,10 @@ app_dockerfile="Dockerfile"
 
 function cleanup() {
   echo "Cleaning up..."
-  docker rm $builder
+  if $(docker ps -a | grep $builder)
+  then
+    docker rm $builder
+  fi
 }
 
 # Always cleanup on exit
@@ -62,10 +65,10 @@ function build_image() {
 
 populate_dockerfile_builder() {
   local builder_file="Dockerfile-builder"
-
-  sed -i "s/^LABEL maintainer=/LABEL maintainer=\"${MAINTAINER}\"" $builder_file
-  sed -i "s/^ENV appname=/ENV appname=${appname}" $builder_file
-  sed -i "s/^ENV appdir=/ENV appdir=${appdir}" $builder_file
+  sed -i "s/^FROM.*$/FROM ${GOLANG_IMAGE}/" $builder_file
+  sed -i "s/^LABEL maintainer=.*$/LABEL maintainer=\"${MAINTAINER}\"/" $builder_file
+  sed -i "s/^ENV appname=.*$/ENV appname=${appname}/" $builder_file
+  sed -i "s|^ENV appdir=.*$|ENV appdir=${appdir}|" $builder_file
 
   echo "Building Go binary with:"
   cat $builder_file
@@ -73,8 +76,8 @@ populate_dockerfile_builder() {
 
 populate_dockerfile() {
   local docker_file="Dockerfile" 
-  sed -i "s/^LABEL maintainer=/LABEL maintainer=\"${MAINTAINER}\"" $docker_file
-  sed -i "s/^CMD/CMD [ \"/${appname}\" ]" $docker_file
+  sed -i "s/^LABEL maintainer=.*$/LABEL maintainer=\"${MAINTAINER}\"/" $docker_file
+  sed -i "s|^CMD.*$|CMD [ \"/${appname}\" ]|" $docker_file
 
   echo "Building image with:"
   cat $docker_file
@@ -83,10 +86,10 @@ populate_dockerfile() {
 main() {
 
   populate_dockerfile_builder
-  #build_binary 
+  build_binary
 
   populate_dockerfile
-  #build_image
+  build_image
 }
 
 main "$@"
